Update README.md
I meant to add this earlier. Maybe I even had it at one point and accidentally removed it? But I think it's more likely that I took a note to add it, and lost track of the note. Anyways, here it is now! It is complementary/compatible with other resources in the guide, not just because of the Python aspect but because of the PyTorch focus.